requirements.txt
numpy
pandas
requests
python-dotenv
scikit-learn
fastapi
uvicorn
pathlib